fix cyclic requires
