+++ app/api/metadata/helpers.js
@@ -1,0 +1,26 @@
+/* eslint import/prefer-default-export: off */
+import type { runtimeType } from './MetadataProviderInterface';
+
+/**
+ * Convert runtime from minutes to hours
+ *
+ * @param  {number} runtimeInMinutes
+ * @return {object}
+ */
+export function parseRuntimeMinutesToObject(
+  runtimeInMinutes: number
+): runtimeType {
+  const hours = runtimeInMinutes >= 60 ? Math.round(runtimeInMinutes / 60) : 0;
+  const minutes = runtimeInMinutes % 60;
+
+  return {
+    full:
+      hours > 0
+        ? `${hours} ${hours > 1 ? 'hours' : 'hour'}${
+            minutes > 0 ? ` ${minutes} minutes` : ''
+          }`
+        : `${minutes} minutes`,
+    hours,
+    minutes
+  };
+}
