+++ package.json
@@ -192,6 +192,7 @@
     "eslint-plugin-import": "^2.17.2",
     "eslint-plugin-jest": "22.4.1",
     "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-prettier": "^3.0.1",
     "eslint-plugin-promise": "4.1.1",
     "eslint-plugin-react": "^7.12.4",
     "eslint-plugin-testcafe": "^0.2.1",
